Protected obscenities achievement against involuntary triggering.
I initially pondered if a regular expression such as the following might
be better suited to match all occurrences of dirty words but then
considered it fighting against the hills:

    \b(ass|cock ?suck|(mother)?fuck|piss|shit|(balls|cunt|tits)\b)

The nice thing about such a fuzzy matcher is that it also triggers words
such as "shitty," but it is hard to get right.  So for now we resort to
bare mentions of those words.